(a) update VerifyKernelDef to handle output type parameters also.
(b) Remove some redundant type constraints in a few kernels.

Related work items: #483

+++ lotus/core/providers/cpu/math/top_k.cc
@@ -29,7 +29,7 @@ REGISTER_KERNEL(KernelDefBuilder("TopK")
                     .SinceVersion(1)
                     .Provider(LotusIR::kCpuExecutionProvider)
                     .TypeConstraint("T", DataTypeImpl::GetTensorType<float>())
-                    .TypeConstraint("I", DataTypeImpl::GetTensorType<float>()),
+                    .TypeConstraint("I", DataTypeImpl::GetTensorType<int64_t>()),
                 TopK<float>);
 
 static int64_t SizeToDim(size_t k, const vector<int64_t>& dims) {
